requirements.txt
torch==1.9.1
opencv_python==4.5.2.54
torchvision==0.10.0
pytorch_lightning==1.3.8
matplotlib==3.4.2
albumentations==1.0.1
numpy==1.21.0
pandas==1.3.0
ella==3.0.13
jsonargparse[signatures]
Pillow==8.3.1
ipywidgets==7.6.2
torchmetrics==0.4.1
